[PhiN] hotfix

+++ app/src/main/java/com/heartsteel/heartory/service/api/UserAPI.kt
@@ -10,7 +10,7 @@ import retrofit2.http.Path
 interface UserAPI {
 
     @GET("users/{id}")
-    suspend fun getUserById(@Path("id")id: Long): Response<ResponseObject<User>>
+    suspend fun getUserById(@Path("id")id: Int): Response<ResponseObject<User>>
 
     @POST("users")
     suspend fun createUser(user: User): Response<ResponseObject<User>>
